- Moved connect from PluginsContainer to withModulePlugins enhancer. Included common values from state into definition so usage of PluginsContainer on page components become shorter and simpler.
