+++ web/client/containers/Embedded.jsx
@@ -10,17 +10,11 @@ import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import assign from 'object-assign';
-import url from 'url';
-const urlQuery = url.parse(window.location.href, true).query;
-import { getMonitoredState } from '../utils/PluginsUtils';
-import ConfigUtils from '../utils/ConfigUtils';
 
 const PluginsContainer = connect((state) => ({
-    mode: urlQuery.mode || (state.browser && state.browser.mobile ? 'mobile' : 'desktop'),
     pluginsState: assign({}, state && state.controls, state && state.layers && state.layers.settings && {
         layerSettings: state.layers.settings
-    }),
-    monitoredState: getMonitoredState(state, ConfigUtils.getConfigProp('monitorState'))
+    })
 }))(require('../components/plugins/PluginsContainer').default);
 
 class Embedded extends React.Component {
