+++ web/client/containers/HolyGrail.jsx
@@ -8,17 +8,8 @@
 import React from 'react';
 
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import url from 'url';
-const urlQuery = url.parse(window.location.href, true).query;
-import { getMonitoredState } from '../utils/PluginsUtils';
-import ConfigUtils from '../utils/ConfigUtils';
+import PluginsContainer from "../components/plugins/PluginsContainer";
 import BorderLayout from '../components/layout/BorderLayout';
-
-const PluginsContainer = connect((state) => ({
-    mode: urlQuery.mode || (urlQuery.mobile || state.browser && state.browser.mobile ? 'mobile' : 'desktop'),
-    monitoredState: getMonitoredState(state, ConfigUtils.getConfigProp('monitorState'))
-}))(require('../components/plugins/PluginsContainer').default);
 
 class HolyGrail extends React.Component {
     static propTypes = {
