+++ web/client/containers/MapViewer.jsx
@@ -10,22 +10,17 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import assign from 'object-assign';
-import url from 'url';
-const urlQuery = url.parse(window.location.href, true).query;
 
 import ConfigUtils from '../utils/ConfigUtils';
-import { getMonitoredState } from '../utils/PluginsUtils';
 import {compose} from "redux";
 import PluginsContainerComponent from "../components/plugins/PluginsContainer";
 
 const PluginsContainer = compose(
     connect((state) => ({
         statePluginsConfig: state.plugins,
-        mode: urlQuery.mode || state.mode || (state.browser && state.browser.mobile ? 'mobile' : 'desktop'),
         pluginsState: assign({}, state && state.controls, state && state.layers && state.layers.settings && {
             layerSettings: state.layers.settings
-        }),
-        monitoredState: getMonitoredState(state, ConfigUtils.getConfigProp('monitorState'))
+        })
     }))
 )(PluginsContainerComponent);
